wam: auto-enable remote debug when agl-devel is on
Apart from preventing remote debugging when agl-devel is not set,
which was tackled in SPEC-2327, we are interested in having remote
debugging available by default when agl-devel is set.

WAM checks if files:
 * /var/agl-devel/preferences/devmode_enabled
 * /var/agl-devel/preferences/debug_system_apps
exist to enable the dev mode and the remote web inspector, so
we create those files by default when agl-devel is set.

Bug-AGL: SPEC-2782

Change-Id: I557a633802e6baa15339708b00719d66adbcf1ac
Signed-off-by: Antia Puentes <apuentes@igalia.com>