meta-pipewire: add recipe to build the bluez-alsa-pipewire gstreamer helper
Unfortunately, the bluez-alsa PCM plugin does not work correctly
when it is used through pipewire (or gstreamer, or anywhere really...).
For this reason I have built a helper client that uses GStreamer
to glue together the bluez-alsa sockets with pipewire.

This helper is implemented as a patch to bluez-alsa so that it can
use its internal private API. In the future this needs some re-thinking

The helper is meant to run in the background as a service and it will
create the appropriate streams in pipewire when it detects a new
device on the bluealsa d-bus interface. Currently it only supports
a2dp-sink and hfp modes (i.e. media player from a phone + calls).
Bluetooth speakers need further policy work in wireplumber that is
too complex to support on the current halibut version of
pipewire/wireplumber.

Bug-AGL: SPEC-2792

Change-Id: Ia0a0c4741dc6f28958e911436edde17ebde1a434
Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>