alexa-voiceagent-service: support building in wakeword engine
Add support for building the "amazonlite" wakeword engine into the Alexa
voiceagent.  A new feature template, agl-voiceagent-alexa-wakeword, is
added to pull in the extra layers/recipes from the additional wakeword
engine files for the SDK that Amazon provides.  The feature pulls in the
existing agl-voiceagent-alexa feature as a dependency to simplify
configuring a build.  The existing patch to work around the Alexa SDK
build's use of AAC_PREFIX as prefix for all packages has been updated to
handle the pryon-lite wakeword engine library.

Bug-AGL: SPEC-3077

Change-Id: I5549ea9000c6713e6becb191f4dcd957ebde18ac
Signed-off-by: Scott Murray <scott.murray@konsulko.com>