alexa-voiceagent-service: Add patch to fix event argument JSON
In several places in the alexa-voiceagent-service code, strings
containing JSON document fragments are put whole into a json-c string
object which is added to the event argument json_object, rather than
being properly added as built up json-c object hierarchies. The result
is an embedded JSON document in a string with extra escaping, which is
not usable on the receiving side without knowing that part of the
event object effectively needs to be run through a JSON parser to be
usable. This is contrary to the voiceagent event documentation and
inconvenient for client implementors, so fix it by tokenizing the
internal JSON payload string into a json_object tree and passing that
as the event argument where necessary.

Note that it is ATM not clear if all affected event argument payloads
are correct, e.g. LocalMediaSource may need some more work.

Bug-AGL: SPEC-3084

Change-Id: I77c97d7eb241d6ccbd98fb379b88336494ce12e9
Signed-off-by: Scott Murray <scott.murray@konsulko.com>