Convert to new override syntax
This is effectively a manual cherry-pick of commit:

37d3b1b316470fdcfebc59d5522bfbfefc025ff6

from the master branch with additional review and fixups.

The intent of these changes is to minimize the effort to backport
fixes from the master branch, which has been updated in preparation
for the switch to the next Yocto LTS release in early 2022.

Bug-AGL: SPEC-4144

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ic649d942ac043edc52976a3d4190590bb1c125e7