aac-module-system-audio: Update PipeWire gstreamer plugins
Update PipeWire sink and source to handle drop of pwaudiosink and
pwaudiosrc in favor of pipewire* replacements.  However, testing
has found that Alexa does not work with pipewiresrc (input hangs
after first interaction), so for now use "alsasrc device=pipewire"
as a workaround as recommended by upstream.  This should be
revisited when PipeWire is next updated.

Bug-AGL: SPEC-4058

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I5f0ebbe80bbafa656dd8e6c542bab666cb481395