meta-agl-flutter: Add option agl-shell plugin patch to embedder
Add a patch to flutter-auto to add an optional agl_shell static
plugin to expose the activate_app call from the agl-shell Wayland
protocol that is used when running against the AGL compositor.
This provides a way for a homescreen application to activate other
application surfaces to switch between multiple applications as is
needed for the AGL demo.  A new PACKAGECONFIG option is defined to
control enabling the plugin, and it is enabled by default in our
build.

Upstreaming this will be discussed with the Toyota team.  Given the
previous discussion about not picking up significant meta-flutter
updates until after CES, it may be that the patch here is carried
for a while.

Bug-AGL: SPEC-4612

Change-Id: Ie8e06c88bbf082ee7dc5126dfb3b24b1bb8deabe
Signed-off-by: Scott Murray <scott.murray@konsulko.com>