ic-container: qemu: Set connector name for IVI layer manaager
The ivi-layer manager configuration file needs to know the name
of display connector to set up the layer configuration.

Adjust the connector name in the configuration file to reflect
the connector names in us on the platform being built.  This
currently only applies to qemu, but could be applied to other
build configurations in future.

Bug-AGL: SPEC-4464

Change-Id: Ie52c9b8639a2bd5c83c4686ff9b557d05a2a7e1e
Signed-off-by: Damian Hobson-Garcia <dhobsong@igel.co.jp>