meta-agl-flutter: fix flutter-gallery build
Add a SRCREV override for flutter-gallery to pick up the version
that was used with Flutter 3.3.x in meta-flutter. This matches our
override of FLUTTER_SDK_TAG to use Flutter 3.3.7, and fixes the
build of it and hence agl-image-flutter.  The override has been
added in agl-flutter.inc to be visible beside the Flutter version
override so they can be updated/removed at the same time easily.

Bug-AGL: SPEC-4819

Change-Id: I27207baa78a498d9e52c21f0b093a3d3bb671f37
Signed-off-by: Scott Murray <scott.murray@konsulko.com>