Add initial DRM lease test images
Add guest and host image as well as associated configuration recipes
for simple DRM lease testing.  The agl-host-image-drm-least-test image
has two guests that each run Weston + desktop-shell with a different
background color.  A known issue at the moment is that networking is
not functional in the guests, this will be investigated further.

Bug-AGL: SPEC-4955

Change-Id: Ib69e627bb251180c2d7a1175de53d2fb1561aa61
Signed-off-by: Scott Murray <scott.murray@konsulko.com>