Add support wl_output version 4 and agl_shell v8
This adds support for wl_output version 4 that allows to match
an output name to a string that can be passed using gRPC APIs.

With this it allows events being sent out by the compositor which
the shell client can react upon.

This can be seen as an intermediary step until flutter-homescreen from
ICS gains (direct) support for gRPC to handle activation that way.

Bug-AGL: SPEC-5004
Change-Id: I143c9dbbc044720c3dee642177d3ae175bfa9a75
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>