Move cluster core package to cluster demo image
This patch is a workaround.
Current guest-image-minimal is used by guest-image-cluster-demo and
guest-image-ivi-demo.  The cluster core package is include at
guest-image-minimal.  In this case, cluster core package is install
to cluster and ivi image.  That cause twin cluster service run on
one system.

This patch fix this issue, but it's only a workaronud.
Will be re-structure for image-recipes.

Bug-AGL: SPEC-4990

Change-Id: I5b9a23042e0796b49ed7703b5860c8a600007251
Signed-off-by: Naoto Yamaguchi <naoto.yamaguchi@aisin.co.jp>